+++ src/fhir/R4R5-common/sdc-export.js
@@ -296,8 +296,8 @@ var self = {
 
       if (answer.score !== null && answer.score !== undefined) {
         ext.push({
-          url: "http://hl7.org/fhir/StructureDefinition/ordinalValue",
-          valueDecimal: parseFloat(answer.score),
+          url: this.fhirExtUrlOptionScore,
+          valueDecimal: answer.score,
         });
       }
       if (ext.length > 0) {